dom/base/test/test_bug518104.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=518104
-->
<head>
  <title>Test for Bug 518104</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=518104">Mozilla Bug 518104</a>
<p id="display"></p>
<script class="testbody" type="text/javascript">

/** Test for Bug 518104 **/
SimpleTest.waitForExplicitFinish();

function done() {
  var iframe = document.getElementById("iframe");
  var divs = iframe.contentWindow.document.getElementsByTagName("p").length;
  is(divs, 0, "<p> got written.")
  SimpleTest.finish();
}

</script>
<div id="content" style="display: none">
  <iframe id='iframe' src="data:text/html,
    <div></div><div></div>
    <script defer src='data:application/javascript,document.write(%2522<p></p>%2522);parent.done();document.close();'></script>">
  </iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
